pyproject.toml
[project]
name = "dataset_crew"
version = "0.1.0"
description = "Language-specific text dataset collection crew using crewAI"
authors = [{ name = "Your Name", email = "you@example.com" }]
requires-python = ">=3.10,<3.13"
dependencies = [
    "crewai[tools]>=0.102.0,<1.0.0",
    "requests>=2.31.0",
    "pandas>=2.0.0",
    "argparse>=1.4.0",
    "serper-dev>=0.1.4",
    "python-dotenv>=1.0.0"
]

[project.scripts]
dataset_crew = "dataset_crew.main:run"
run_crew = "dataset_crew.main:run"
train = "dataset_crew.main:train"
replay = "dataset_crew.main:replay"
test = "dataset_crew.main:test"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.crewai]
type = "crew"
